requirements.txt
django
psycopg2
djangorestframework
psycopg2
django-cors-headers
djangorestframework-simplejwt